Move drag & drop support logic to chooser component
